initial commit of cleaning object creation process
move `global` variable mehtod to `glbl.py` and hand corresponding objects to other modules
