+++ xpd_profile_file/0-init_simulation.py
@@ -1,0 +1,5 @@
+from simulator.areadetector import pe1c
+from simulator.shutter import shctl1
+
+
+
